refactor(front: user, user following): ユーザー詳細とフォローしているロースター一覧ページのリファクタリング
