+++ front/src/features/users/components/pages/User.tsx
@@ -1,39 +1,22 @@
 import type { FC } from 'react'
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { useParams, useNavigate } from 'react-router-dom'
 
 import { ContentHeader } from '@/components/Elements/Header'
 import { ContentHeaderTitle } from '@/components/Elements/Header/ContentHeaderTitle'
 import { Spinner } from '@/components/Elements/Spinner'
-import { useAuth } from '@/features/auth'
-import { getUser } from '@/features/users/api/getUser'
 import { UserCard } from '@/features/users/components/organisms/UserCard'
-import type { User as UserType } from '@/features/users/types'
-import { useMessage } from '@/hooks/useMessage'
+import { useGetUser } from '@/features/users/hooks/useGetUser'
 
 export const User: FC = () => {
-  const [user, setUser] = useState<UserType>()
-  const [loading, setLoading] = useState(false)
   const urlParams = useParams<{ id: string }>()
-  const { authHeaders } = useAuth()
-  const { showMessage } = useMessage()
   const navigate = useNavigate()
+  const { user, getUser, loading } = useGetUser()
 
   useEffect(() => {
     // urlParams.idが数値かどうか評価
     if (urlParams.id && !Number.isNaN(parseInt(urlParams.id, 10))) {
-      setLoading(true)
-      getUser({ headers: authHeaders, id: urlParams.id })
-        .then((response) => {
-          setUser(response.data)
-        })
-        .catch(() => {
-          navigate('/')
-          showMessage({ message: 'ユーザーが存在しません', type: 'error' })
-        })
-        .finally(() => {
-          setLoading(false)
-        })
+      getUser(urlParams.id)
     } else {
       navigate('/')
     }
