+++ front/src/features/users/hooks/useGetUser.ts
@@ -1,0 +1,33 @@
+import { useCallback, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
+import { useAuth } from '@/features/auth'
+import type { User } from '@/features/users'
+import { getUser as getUserRequest } from '@/features/users/api/getUser'
+import { useMessage } from '@/hooks/useMessage'
+
+export const useGetUser = () => {
+  const navigate = useNavigate()
+  const { showMessage } = useMessage()
+  const { authHeaders } = useAuth()
+
+  const [loading, setLoading] = useState(false)
+  const [user, setUser] = useState<User>()
+
+  const getUser = useCallback((id: string) => {
+    setLoading(true)
+    getUserRequest({ headers: authHeaders, id })
+      .then((response) => {
+        setUser(response.data)
+      })
+      .catch(() => {
+        navigate('/')
+        showMessage({ message: 'ユーザーが存在しません', type: 'error' })
+      })
+      .finally(() => {
+        setLoading(false)
+      })
+  }, [])
+
+  return { user, getUser, loading }
+}
